move the test to `/tests/test_buf.rs`
